refactor(settings): 拆出外观域状态 hook 与外观页签容器
- 新增 settings/appearance/：useThemePackages（主题包选择/导入/删除/AI调色）、useBackgroundSettings（全局/终端背景与透明度）、useProgramFonts（程序字体导入与拖拽分配）、useTerminalPreferences（终端显示偏好与关键字规则）、AppearanceTabPane（外观域常驻容器）
- SettingsModal.tsx 1755 -> 1194 行（language 与终端交互类通用设置状态保留在弹窗层，待随通用页签容器迁移）
